Stdout to additional print datetime in human friendly format
